+++ KreaitFirebaseBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Kreait\FirebaseBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class KreaitFirebaseBundle extends Bundle
-{
-}